priceをBigDecimalにする

+++ src/main/java/com/javajo/javajo_jewels/model/Product.java
@@ -21,7 +21,7 @@ public class Product {
                 entity.getId(),
                 entity.getName(),
                 entity.getDescription(),
-                entity.getPrice().intValue(),
+                entity.getPrice(),
                 entity.getImageUrl()
         );
     }
